[Kernel] Resolve sixth batch of reviews of adding properties
